2006-09-21  Sancho Lerena  <slerena@artica.es>
	* pandoradb.sql: Altered tagente_datos field "datos" to have
	DOUBLE(18,2) instead of DOUBLE (12,2).

	* pandora_db1.1_to1.2.sql: Altered tagente_datos field "datos" to
	have DOUBLE(18,2) instead of DOUBLE (12,2).



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@183 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/pandoradb_1.1_to_1.2.sql
@@ -174,7 +174,7 @@ DROP TABLE tagente_datos;
 CREATE TABLE `tagente_datos` (
   `id_agente_datos` bigint(10) unsigned NOT NULL auto_increment,
   `id_agente_modulo` bigint(4) NOT NULL default '0',
-  `datos` double(12,2) default NULL,
+  `datos` double(18,2) default NULL,
   `timestamp` datetime NOT NULL default '0000-00-00 00:00:00',
   `id_agente` bigint(4) unsigned NOT NULL default '0',
   PRIMARY KEY  (`id_agente_datos`),
